Fix skipping tests if Qdrant server is not running
