+++ tests/integration_tests/vectorstores/qdrant/async/test_from_texts.py
@@ -10,6 +10,13 @@ from tests.integration_tests.vectorstores.fake_embeddings import (
     ConsistentFakeEmbeddings,
 )
 
+from .common import qdrant_is_not_running
+
+# Skipping all the tests in the module if Qdrant is not running on localhost.
+pytestmark = pytest.mark.skipif(
+    qdrant_is_not_running(), reason="Qdrant server is not running"
+)
+
 
 @pytest.mark.asyncio
 async def test_qdrant_from_texts_stores_duplicated_texts() -> None:
